backend/requirements.txt
Django==3.2.7
django-cors-headers==3.8.0
django-filter==2.4.0
djangorestframework==3.12.4
pillow==9.3.0